*** empty log message ***

+++ include/libtorrent/peer_connection.hpp
@@ -86,7 +86,7 @@ namespace libtorrent
 		inline T read_impl(InIt& start, type<T>)
 		{
 			T ret = 0;
-			for (int i = 0; i < sizeof(T); ++i)
+			for (int i = 0; i < (int)sizeof(T); ++i)
 			{
 				ret <<= 8;
 				ret |= static_cast<unsigned char>(*start);
@@ -98,7 +98,7 @@ namespace libtorrent
 		template <class T, class OutIt>
 		inline void write_impl(T val, OutIt& start)
 		{
-			for (int i = sizeof(T)-1; i >= 0; --i)
+			for (int i = (int)sizeof(T)-1; i >= 0; --i)
 			{
 				*start = static_cast<unsigned char>((val >> (i * 8)) & 0xff);
 				++start;
